split nmod_vec.h -> nmod.h, nmod_vec.h and document a couple of things
