+++ doc/source/index.rst
@@ -89,6 +89,7 @@ Integers mod n
 .. toctree::
    :maxdepth: 1
 
+   nmod.rst
    nmod_vec.rst
    nmod_mat.rst
    nmod_poly.rst
